[#27] chore: Move surveysPerPage to a separate constants folder
